+++ src/meetup/MeetupLogo.tsx
@@ -1,0 +1,53 @@
+import {AbsoluteFill, spring, useCurrentFrame, useVideoConfig} from 'remotion';
+import {LyonJSLogo} from '../components/LyonJSLogo';
+
+export const MeetupLogo: React.FC<{endAnimationShift: number}> = ({
+	endAnimationShift,
+}) => {
+	const frame = useCurrentFrame();
+	const {fps} = useVideoConfig();
+
+	const scaleUpLogo = spring({
+		frame,
+		from: 0,
+		to: 1,
+		fps,
+		durationInFrames: 30,
+	});
+	const scaleDownLogo = spring({
+		frame: frame - endAnimationShift,
+		from: 0,
+		to: 1,
+		fps,
+		config: {
+			mass: 0.6,
+			damping: 5,
+			stiffness: 50,
+		},
+		durationInFrames: 65,
+	});
+
+	const opacityUp = spring({frame, from: 0, to: 1, fps, durationInFrames: 50});
+	const opacityDown = spring({
+		frame: frame - endAnimationShift,
+		from: 0,
+		to: 0.5,
+		fps,
+		durationInFrames: 70,
+	});
+
+	return (
+		<AbsoluteFill style={{display: 'flex', alignItems: 'center'}}>
+			<LyonJSLogo
+				style={{
+					position: 'absolute',
+					top: 150,
+					height: 400,
+					transform: `scale(${Math.abs(scaleUpLogo - scaleDownLogo)})`,
+					filter: 'drop-shadow(0px 0px 15px #000000)',
+					opacity: opacityUp - opacityDown,
+				}}
+			/>
+		</AbsoluteFill>
+	);
+};
